% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.data.R
\name{clean.data}
\alias{clean.data}
\title{Clean Raw Player Tracking Data}
\usage{
clean.data(
  player.speed,
  satellite.quality,
  max.speed = 12,
  max.satellite.quality = 2,
  metrics = "m/s"
)
}
\arguments{
\item{player.speed}{player speed vector}

\item{satellite.quality}{typically refers the hdop vector}

\item{max.speed}{player's max attainable speed, default set to 12 m/s}

\item{max.satellite.quality}{max hdop quality, default set to 2}

\item{metrics}{metric or unit of the player speed vector}
}
\description{
A data set that contains a player's speed, player's speed in metric, and satellite quality. Data only includes realistic 
and attainable player speeds and high satellite quality.
}
\details{
This function is intended to help users clean and process their data. In particular, `clean.data` removes unrealistic player 
running values (default set to > 12 m/s), maintains high satellite quality (hdop < 2), and can convert a player's speed 
to `m/s`. 

Please note that this function simply removes observations that do not meet inclusion criterion. Users are encouraged 
to visualize their data prior to processing it to ensure that the tracking data is realistic. For example, GPS units 
are often left on during bus rides and at half-time. In these cases, users should create distinct data sets that 
correspond to a player's actual time while wearing the GPS units before processing and modelling the data.

If the player tracking data does not contain satellite quality, set `satellite.quality = NA`. 

This function can convert the following units to m/s: feet (`ft`), yards (yd), kilometers (km), and miles (mi) per 
second (s), minute (min), and hour (h). For example, a player's speed in yards per second should be input as `yd/s`.
}
