% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critspeed.R
\name{compile}
\alias{compile}
\title{Greatest Maximal Mean and Median Speed per Duration}
\usage{
compile(data, ...)
}
\arguments{
\item{data}{data set containing at least two variables (see notes)}

\item{...}{other data sets that fulfill the requirements described in **Notes**}
}
\description{
A lightweight function that returns the individuals greatest max mean and median speed per duration.
}
\details{
This function provides practitioners the ability to return the greatest max mean and median speed per duration. The `data`
and `...` arguments must be `data.frame`s or `data.table`s. It is suggested that practitioners use the `critspeed()` function 
to generate the required `data.frame`s. 

Users are returned a `data.frame` that consists of either two or three variables. The first variable is always `duration`. The 
second variable is either the maximal mean speed or maximal median speed per duration. If users input `data.frame`s that include 
maximal mean and maximal median speeds, they will be returned a data frame with `duration`, `max.mean.speed`, and `max.median.speed` 
variables (if they generated the data using the `critspeed()` function). This ensures that max mean and max median 
speeds are not assumed to be interchangeable.
}
\note{
Users must be aware of the following requirements for the following arguments:    

\enumerate{
\item `data` is a `data.frame` must satisfy the following criteria:
\itemize{
\item The first column is the duration variable
\item The second column is the maximal mean speed or maximal median speed variable
\item Subsequent columns will be return an error and must be removed before processing
}

Users can generate appropriate `data.frame`s using the `critspeed()` function

}
}
