% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs.results.model.R
\name{crit.speed.results}
\alias{crit.speed.results}
\title{Critical Speed Model Results}
\usage{
crit.speed.results(
  player.speed,
  global.data = FALSE,
  sample.rate = 10,
  cv.2 = 120,
  dur = 600,
  d.prime.estim = 100,
  crit.speed.estim = 3.5,
  max.speed.estim = 12
)
}
\arguments{
\item{player.speed}{speed column}

\item{global.data}{default set to FALSE}

\item{sample.rate}{in Hz}

\item{cv.2}{min duration for CV2. Default set to 120 s}

\item{dur}{duration}

\item{d.prime.estim}{default set to 100}

\item{crit.speed.estim}{default set to 3.5}

\item{max.speed.estim}{default set to 12}
}
\description{
Critical Speed Model Results
}
