% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.results.compare.R
\name{critspeed.list}
\alias{critspeed.list}
\title{Build Multiple critspeed Data Sets}
\usage{
critspeed.list(
  data = list(),
  data2 = list(),
  speed.col = "max.mean.speed",
  raw.data = TRUE,
  sample.rate = 10,
  dur = 600,
  roecker = FALSE
)
}
\arguments{
\item{data}{group 1 of sessions}

\item{data2}{group 2 of sessions}

\item{speed.col}{column name for player speed}

\item{raw.data}{data type being either raw (unprocessed) or critspeed (processed), defaults set to TRUE}

\item{sample.rate}{in Hz, default set to 10 Hz}

\item{dur}{max duration, default set to 600 s}

\item{roecker}{indicates whether multiple durations are provided, default set to FALSE}
}
\description{
Returns a list of data sets that contain the max mean or max median speed per duration.
}
\details{
This function, and `cs.results.observed.list` return the max mean or max median player speed per 
duration. The functions differ in that `cs.results.observed.list` feeds into `model.results.compare` 
which requires alternate variable names. 

For users interested in manually generating multiple critspeed data sets, it is recommended 
that they call `critspeed.list` which will return variable names that correspond with whether 
`roecker` is set to `TRUE` or `FALSE`.
}
