% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs.results.fitted.R
\name{cs.results.fitted}
\alias{cs.results.fitted}
\title{Fitted Results}
\usage{
cs.results.fitted(
  player.speed,
  global.data = FALSE,
  sample.rate = 10,
  cv.2 = 120,
  dur = 600,
  d.prime.estim = 100,
  crit.speed.estim = 3.5,
  max.speed.estim = 12
)
}
\arguments{
\item{player.speed}{speed column}

\item{global.data}{default set to FALSE}

\item{sample.rate}{in Hz}

\item{cv.2}{start duration for CV2}

\item{dur}{duration}

\item{d.prime.estim}{default set to 100}

\item{crit.speed.estim}{default set to 3.5}

\item{max.speed.estim}{default set to 12}
}
\description{
Fitted Results
}
