% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs.results.observed.list.R
\name{cs.results.observed.list}
\alias{cs.results.observed.list}
\title{Critical Speed List Results}
\usage{
cs.results.observed.list(
  data = list(),
  data2 = list(),
  speed.col = "speed.col.name",
  global.data = FALSE,
  sample.rate = 10,
  dur = 600
)
}
\arguments{
\item{data}{group 1 of sessions}

\item{data2}{group 2 of sessions}

\item{speed.col}{column name for player speed}

\item{global.data}{default set to FALSE}

\item{sample.rate}{in Hz}

\item{dur}{duration}
}
\description{
Critical Speed List Results
}
