% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs.results.plot.R
\name{cs.results.plot}
\alias{cs.results.plot}
\title{Fitted Plots}
\usage{
cs.results.plot(
  player.speed,
  global.data = FALSE,
  sample.rate = 10,
  dur = 600,
  cv.2 = 120,
  d.prime.estim = 150,
  crit.speed.estim = 3.5,
  max.speed.estim = 12,
  log.dur = FALSE,
  resid = FALSE
)
}
\arguments{
\item{player.speed}{speed column}

\item{global.data}{default set to FALSE}

\item{sample.rate}{in Hz}

\item{dur}{duration}

\item{cv.2}{min duration for CV2}

\item{d.prime.estim}{default set to 100}

\item{crit.speed.estim}{default set to 3.5}

\item{max.speed.estim}{default set to 12}

\item{log.dur}{default set to FALSE. Takes log base 10 of duration}

\item{resid}{residual plots. Not yet implemented}
}
\description{
Fitted Plots
}
