% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.model.R
\name{fitted.model}
\alias{fitted.model}
\title{Fitted Model Values}
\usage{
\method{fitted}{model}(
  data,
  model = c("two.param", "three.param", "five.param", "omvd"),
  cv.2 = 120,
  lower = c(cs = 1, dp = 10, v0 = 5.8, a = NA, b = NA, f = NA),
  upper = c(cs = 5.8, dp = 800, v0 = 12, a = NA, b = NA, f = NA),
  start = c(cs = 4, dp = 150, v0 = 9, a = -2.71, b = 2.195, f = 0.2)
)
}
\arguments{
\item{data}{data set containing at least two variables (see notes)}

\item{model}{critical speed model name as either `two.param`, `three.param`, `five.param`, or `omvd`}

\item{cv.2}{minimum duration to fit the two parameter function, default set to 120 s (see notes)}

\item{lower}{lower bounds for parameter fits (see notes)}

\item{upper}{upper bounds for parameter fits (see notes)}

\item{start}{starting values for parameter fits (see notes)}
}
\description{
A `data.frame` that contains the fitted model values.
}
\details{
`fitted.model()` is similar to the `stats` `fitted()` and `residuals()` functions. This function is more complex in that 
it returns the original data, and the fitted and residual values for the selected models. This gives users the ability to generate 
`data.frame`s that best suit their goals. 

The `data.frame` that is returned will always include `duration`, `log.duration`, and `max.speed` variables. It was decided that `max.speed` should 
represent maximal mean speed and maximal median speed. Discriminating between the two is impractical for the intended uses of this function.
}
\note{
Users must be aware of the following requirements for the following arguments:    

\enumerate{
\item `data` is a `data.frame` that must satisfy the following criteria:
\itemize{
\item The first column is the duration variable
\item The second column is the maximal mean speed or maximal median speed variable
\item Subsequent columns will be disregarded
}

Users can generate appropriate `data.frame`s using the `critspeed()` function

\item `cv.2` is the minimum duration used to fit the two parameter model. The default is set to 120 s as suggested by the prevailing 
literature.

\item `lower`, `upper`, and `start` are the lower and upper bounds for the parameter fits, and initial parameter values, respectively 
The default values are set based on various considerations including the prevailing literature and trial-and-error. The defaults are 
outlined below in the form of `parameter: lower bound, start value, upper bound`:
\itemize{
\item d prime: 10, 150, 800 in meters
\item critical speed: 1, 4, 5.8 in meters per second
\item maximal speed: 4, 9, 12 in meters per second
\item a: NA, -2.71, NA and is unitless
\item b: NA, 2.195, NA and is unitless
\item f: NA, 0.2, NA and is unitless
}

}
}
\seealso{
[model], [model.parameters], [critspeed]
}
