% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max.mean.speed.R
\name{max.mean.speed}
\alias{max.mean.speed}
\title{Max Mean Speed}
\usage{
\method{max}{mean.speed}(x, n = 600)
}
\arguments{
\item{x}{player speed vector}

\item{n}{duration}
}
\description{
This is a helper function that feeds into `max.mean.speed.df`.
}
\details{
The arguments for this function are minimal and it is not suggested that this function be 
used on its own.

`n` is in arbitrary units and does not take into account sample rate. Therefore, 
users should be cautious when calling `max.mean.speed` exclusively.
}
