% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max.median.speed.R
\name{max.median.speed.df}
\alias{max.median.speed.df}
\title{Max Median Speed Data Frame}
\usage{
\method{max}{median.speed.df}(player.speed, sample.rate = 10, dur = "roecker")
}
\arguments{
\item{player.speed}{player speed vector}

\item{sample.rate}{in Hz, default set to 10 Hz}

\item{dur}{duration(s), default set to "roecker"}
}
\description{
A data frame that contains the greatest median speed per duration.
}
\details{
This function emulates the maximal median speed calculation that was introduced 
by Roecker, Mahler, Heyde, Röll, & Gollhofer (2017).

This function calls on `max.median.speed` to calculate the maximal median 
speed that a player can maintain for a given duration. The `dur` argument can take on 
either a value of 1L (a single value), a vector of multiple values,or "roecker". 
When users input a value of 1L, the functions treats it as the maximal duration that 
the maximal median speed should be calculated per duration. For example, if `dur` is set 
to 600 with `sample.rate` set to 10 Hz, a data frame will be returned with durations 
ranging from 0.1 to 600 s. If a vector of values is provided to the `dur` argument, the 
function will call upon these values in seconds. Therefore, a `dur` of c(0.1, 10, 100, 600) 
returns a data frame containing four observations of 0.1, 10, 100, and 600 s. Finally, if 
`dur` is set to "roecker", durations are set to those suggested in the original paper 
by Roecker, Mahler, Heyde, Röll, & Gollhofer (2017): 0.3, 0.5, 1, 2, 3, 4, 5, 6.5, 10, 
13.5, 18, 30, 60, 120, 300, 600, 900, 1200, 1800, 2400, and 2700 s.

The max median speed is calculated using `Rfast::med` which uses a C++ back-end. 
In doing so, this function can be temporarily taxing on your system. It is suggested that this 
function be called when the system has available RAM and CPU. Otherwise, an error message can be 
returned indicating that a vector of size x Mb cannot be assigned.

Please note that calculating the max median value can be significantly slower 
than calculating the max mean speed. For those interested in the max mean speed, please 
see `max.mean.speed.df`.
}
\references{
Roecker, K., Mahler, H., Heyde, C., Röll, M., & Gollhofer, A. (2017). 
The relationship between movement speed and duration during soccer matches. PLoS 
ONE, 12(7), e0181781. https://doi.org/10.1371/journal.pone.0181781
}
\seealso{
max.median.speed, max.mean.speed.df, max.mean.speed
}
