% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.fits.plot.R
\name{model.fits.plot}
\alias{model.fits.plot}
\title{Model Fits Plots}
\usage{
model.fits.plot(
  dur = 600,
  log.dur = FALSE,
  model.coefs,
  model.fits,
  roecker = FALSE
)
}
\arguments{
\item{dur}{max duration, default set to 600 s}

\item{log.dur}{should the fitted plots be returned on a log10 scale, default set to FALSE}

\item{model.coefs}{data set returned when calling critspeed::model.results}

\item{model.fits}{data set returned when calling critspeed::model.fits}

\item{roecker}{indicates whether multiple durations are provided, default set to FALSE}
}
\description{
This function is a helper function for `model.plots` when `plots = "fit"`.
}
\details{
When calling this function explicitly, the argument `model.fits` and `model.coefs` takes on the data frame that 
is returned when calling `model.fits` and `model.results`, respectively.
}
