% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.qq.plot.R
\name{model.qq.plot}
\alias{model.qq.plot}
\title{Model Q-Q Plots}
\usage{
model.qq.plot(model.fits)
}
\arguments{
\item{model.fits}{data set returned when calling critspeed::model.fits}
}
\description{
This function is a helper function for `model.plots` when `plots = "qq"`.
}
\details{
When calling this function explicitly, the argument `model.fits` takes on the data frame that 
is returned when calling `model.fits`.
}
