% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.residuals.plot.R
\name{model.residuals.plot}
\alias{model.residuals.plot}
\title{Model Residuals Plots}
\usage{
model.residuals.plot(dur = 600, log.dur = FALSE, model.fits, roecker = FALSE)
}
\arguments{
\item{dur}{max duration, default set to 600 s}

\item{log.dur}{should the fitted plots be returned on a log10 scale, default set to FALSE}

\item{model.fits}{data set returned when calling critspeed::model.fits}

\item{roecker}{indicates whether multiple durations are provided, default set to FALSE}
}
\description{
A residuals plot for the 2P, 3P, 5P, and OmVD models.
}
\details{
This function is a helper function for `model.plots` when `plots = "resid"`. 

When calling this function explicitly, the argument `model.fits` takes on the data frame that 
is returned when calling `model.fits`.
}
