% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.results.compare.R
\name{model.results.compare}
\alias{model.results.compare}
\title{Compare Model Parameters and Goodness-of-Fits}
\usage{
model.results.compare(
  data = list(),
  data2 = list(),
  speed.col = "max.mean.speed",
  raw.data = TRUE,
  sample.rate = 10,
  dur = 600,
  roecker = FALSE,
  cv.2 = 120,
  d.prime.estim = 150,
  max.speed.estim = 12,
  crit.speed.estim = 3.5
)
}
\arguments{
\item{data}{group 1 of sessions}

\item{data2}{group 2 of sessions}

\item{speed.col}{column name for player speed}

\item{raw.data}{data type being either raw (unprocessed) or critspeed (processed), defaults set to TRUE}

\item{sample.rate}{in Hz, default set to 10 Hz}

\item{dur}{max duration, default set to 600 s}

\item{roecker}{indicates whether multiple durations are provided, default set to FALSE}

\item{cv.2}{start duration for CV2, default set to 120 s}

\item{d.prime.estim}{D' estimate, default set to 100 m}

\item{max.speed.estim}{max speed, default set to 12 m/s}

\item{crit.speed.estim}{critical speed estimate, default set to 3.5 m/s}
}
\description{
A data set with model results from multiple player sessions.
}
\details{
This function allows users to compare model results from multiple player sessions. `data` and `data2` can also take on 
one or multiple sessions. The advantage of `data` and `data2` taking on a list of sessions allows users to input 
a week or month of sessions and compare parameter estimates between weeks or months. Doing so provides practitioners 
the ability to track changes in a player's critical speed, D', and max speed. 

If `raw.data = TRUE` function calls on `critspeed::critspeed` to process the data which it then feeds into `model.results`. 
If `raw.data = FALSE`, `model.results` is called directly. 

As with `critspeed::critspeed`, this function relies heavily on `data.table` and `Rfast`, both of which use a C++ 
backend. In doing so, processing time is significantly decreased at the cost of computer resources. It is suggested that 
CPU and RAM be available when running `critspeed::critspeed` to minimize the risk of an error occuring. Common `Error` messages 
returned are that there is not enough memory available to assign a vector of size x MB. To avoid this error, 
sufficient RAM must be available on your local machine.
}
\seealso{
critspeed, max.mean.speed.df, max.median.speed.df, model.results

critspeed.list
}
