% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.plots.R
\name{plot.model}
\alias{plot.model}
\title{Plots for Model Fits, Residuals, and Normality}
\usage{
\method{plot}{model}(
  data,
  model = c("two.param", "three.param", "five.param", "omvd"),
  output = c("summary", "fit", "residual", "qq", "density"),
  cv.2 = 120,
  log.dur = FALSE,
  method = c("mean", "median"),
  lower = c(cs = 1, dp = 10, v0 = 5.8, a = NA, b = NA, f = NA),
  upper = c(cs = 5.8, dp = 800, v0 = 12, a = NA, b = NA, f = NA),
  start = c(cs = 4, dp = 150, v0 = 9, a = -2.71, b = 2.195, f = 0.2),
  ...
)
}
\arguments{
\item{data}{data set containing at least two variables (see notes)}

\item{model}{critical speed model name as either `two.param`, `three.param`, `five.param`, or `omvd`}

\item{output}{the type of plot returned, default is set to `summary`}

\item{cv.2}{minimum duration to fit the two parameter function, default set to 120 s (see notes)}

\item{log.dur}{plot duration as log10(duration), default set to FALSE}

\item{method}{indicates whether the max mean or max median speed were calculated}

\item{lower}{lower bounds for parameter fits (see notes)}

\item{upper}{upper bounds for parameter fits (see notes)}

\item{start}{starting values for parameter fits (see notes)}

\item{...}{supplement arguments passed, used primarily for developmental purposes and should not be assigned}
}
\description{
Provides the user the ability to produce plots that can return model fits, residuals, and normality.
}
\details{
This function returns the following plots: 
\itemize{
\item `summary`: a four panel plot that includes fitted values, residuals vs fitted, qq, and density of residuals plots
\item `fit`: observed and fitted values, including a horizontal asymptote with a critical speed annotation
\item `residual`: residual vs fitted values to observe scedacity
\item `qq`: a Q-Q plot for observation distribution
\item `density`: density of residuals to observe whether residuals are normally distributed
}
}
\note{
Users must be aware of the following requirements for the following arguments:    

\enumerate{
\item `data` is a `data.frame` must satisfy the following criteria:
\itemize{
\item The first column is the duration variable
\item The second column is the maximal mean speed or maximal median speed variable
\item Subsequent columns will be disregarded
}

Users can generate appropriate `data.frame`s using the `critspeed()` function

\item `cv.2` is the minimum duration used to fit the two parameter model. The default is set to 120 s as suggested by the prevailing 
literature.

\item `lower`, `upper`, and `start` are the lower and upper bounds for the parameter fits, and initial parameter values, respectively 
The default values are set based on various considerations including the prevailing literature and trial-and-error. The defaults are 
outlined below in the form of `parameter: lower bound, start value, upper bound`:
\itemize{
\item d prime: 10, 150, 800 in meters
\item critical speed: 1, 4, 5.8 in meters per second
\item maximal speed: 4, 9, 12 in meters per second
\item a: NA, -2.71, NA and is unitless
\item b: NA, 2.195, NA and is unitless
\item f: NA, 0.2, NA and is unitless
}

}
}
\seealso{
[model.parameters], [model], [fitted.model]
}
