% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed.observations.r
\name{speed.observations}
\alias{speed.observations}
\title{Max Mean and Median Observations}
\usage{
speed.observations(
  player.speed,
  ...,
  dur = NULL,
  sample.rate = 10,
  method = c("mean", "median")
)
}
\arguments{
\item{player.speed}{the player's speed vector}

\item{...}{supplemental speed vectors}

\item{dur}{duration of length 1L to calculate max mean or max median speed}

\item{sample.rate}{in Hz, default set to 10 Hz}

\item{method}{indicates whether the max mean or max median speed should be calculated}
}
\description{
A \code{data.frame} that contains the observations that underpin the maximal mean or median speed.
}
\details{
The \code{speed.observations()} function returns a \code{data.frame} that contains three variables: \code{index}, \code{time} and 
\code{player.speed}. The \code{index} variable corresponds to which vector the observations came from, and \code{time} variable 
corresponds to the time, or duration, that the observations correspond to. 

When there is one speed vector input, the default index is \code{1}.
}
\examples{

data(sessionRaw)

# Max Mean Speed Observations

max.mean.observations <- speed.observations(player.speed = sessionRaw$speed,
dur = 20,
sample.rate = 10,
method = "mean")

# head(max.mean.observations)

# Max Median Speed Observations

max.median.observations <- speed.observations(player.speed = sessionRaw$speed,
dur = 20,
sample.rate = 10,
method = "median")

# head(max.median.observations)
}
