% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critspeed.R
\name{compile}
\alias{compile}
\title{Greatest Maximal Mean and Median Speed per Duration}
\usage{
compile(data, ..., index = TRUE)
}
\arguments{
\item{data}{data set containing at least two variables (see notes)}

\item{...}{other data sets that fulfill the requirements described in \strong{Notes}}

\item{index}{returns the session that contained the greatest max mean or max median speed, default set to TRUE}
}
\description{
A lightweight function that returns the individuals greatest max mean and median speed per duration.
}
\details{
This function provides practitioners the ability to return the greatest max mean and median speed per duration. The \code{data}
and \code{...} arguments must be \code{data.frame}s or \code{data.table}s. It is suggested that practitioners use the \code{critspeed()} function 
to generate the required \code{data.frame}s. 

Users are returned a \code{data.frame} that consists of either two or three variables. The first variable is always \code{duration}. The 
second variable is either the maximal mean speed or maximal median speed per duration. If users input \code{data.frame}s that include 
maximal mean and maximal median speeds, they will be returned a data frame with \code{duration}, \code{max.mean.speed}, and \code{max.median.speed} 
variables (if they generated the data using the \code{critspeed()} function). This ensures that max mean and max median 
speeds are not assumed to be interchangeable.

Indexing returns the session from which the max mean or max median speed was calculated. The variable, `.id`, returns an integer that corresponds with 
the position of the \code{data.frame} when input into \code{compile()}.
}
\note{
Users must be aware of the following requirements for the following arguments:    

\enumerate{
\item \code{data} is a \code{data.frame} must satisfy the following criteria:
\itemize{
\item The first column is the duration variable called \code{duration}
\item The second column is the maximal mean speed or maximal median speed variable and named accordingly (\code{max.mean.speed} or \code{max.median.speed})
\item Subsequent columns will return an error and must be removed before processing
}
\item \code{index} returns a warning when variable names are inconsistent

Users can generate appropriate \code{data.frame}s using the \code{critspeed()} function

}
}
\examples{

data(sessionMaxMeanSpeed)
data(sessionMaxMedianSpeed)
data(sessionRaw)

# Compile Max Mean Speed

\dontrun{

session.mean <- critspeed(sessionRaw$speed,
dur = 600,
sample.rate = 10,
method = "mean")

compile.mean <- compile(sessionMaxMeanSpeed, session.mean, index = TRUE)

# head(compile.mean)


# Compile Max Mean and Max Median Speed

compile.mean.median <- compile(sessionMaxMeanSpeed, session.mean, sessionMaxMedianSpeed, index = TRUE)

# returns a warning and will not index the max mean and max median speeds
# head(compile.mean.median)
}


}
\seealso{
\code{\link{critspeed}}
}
