% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max.mean.speed.R
\name{max.mean.speed.df}
\alias{max.mean.speed.df}
\title{Max Mean Speed Data Frame}
\usage{
\method{max}{mean.speed.df}(player.speed, sample.rate = 10, dur = 600)
}
\arguments{
\item{player.speed}{player speed vector}

\item{sample.rate}{in Hz}

\item{dur}{duration}
}
\description{
Max Mean Speed Data Frame
}
