% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max.mean.speed.R
\name{max.mean.speed}
\alias{max.mean.speed}
\title{Max Mean Speed}
\usage{
\method{max}{mean.speed}(x, n = 600)
}
\arguments{
\item{x}{player speed vector}

\item{n}{duration}
}
\description{
Max Mean Speed
}
