% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max.median.speed.R
\name{max.median.speed.df}
\alias{max.median.speed.df}
\title{Max Median Speed Data Frame}
\usage{
\method{max}{median.speed.df}(player.speed, sample.rate = 10, dur = "roeker")
}
\arguments{
\item{player.speed}{player speed vector}

\item{sample.rate}{in Hz}

\item{dur}{durations}
}
\description{
Max Median Speed Data Frame
}
