% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max.median.speed.R
\name{max.median.speed}
\alias{max.median.speed}
\title{Max Median Speed}
\usage{
\method{max}{median.speed}(x, n = "roeker", sample.rate = 10)
}
\arguments{
\item{x}{player speed vector}

\item{n}{durations}

\item{sample.rate}{in Hz}
}
\description{
Max Median Speed
}
